+++ web/src/components/pages/Content/index.tsx
@@ -1,5 +1,5 @@
 import { AuthenticationRequiredPage } from "@reearth-cms/auth";
-import ContentList from "@reearth-cms/components/organisms/Content/ContentList";
+import ContentList from "@reearth-cms/components/organisms/Project/Content/ContentList";
 
 const ContentPage: React.FC = () => {
   return (
